Refactor metrics into dedicated package with Start() entrypoint, add labeled Prometheus metrics and configurable polling
